pyproject.toml
[tool.poetry]
name = "solidity"
version = "0.1.0"
description = ""
authors = ["mrotame <mrotame@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
eth-wake = "^4.9.0"
python-dotenv = "^1.0.1"
web3 = "^6.19.0"
toml = "^0.10.2"
firebase-admin = "^6.5.0"
requests = "^2.32.2"
loguru = "^0.7.2"
rdoclient = "^1.5.0"
six = "^1.16.0"


[tool.poetry.group.dev.dependencies]
eth-wake = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
